lvl parser reorg
