+++ fast64_internal/oot/scene/importer/__init__.py
@@ -1,0 +1,1 @@
+from .scene import parseScene
