+++ fast64_internal/oot/scene/operators.py
@@ -10,7 +10,7 @@ from ...utility import PluginError, raisePluginError, ootGetSceneOrRoomHeader
 from ..oot_utility import ExportInfo, sceneNameFromID, getEnumName
 from ..oot_level_writer import ootExportSceneToC
 from ..oot_constants import ootEnumMusicSeq, ootEnumSceneID
-from ..oot_level_parser import parseScene
+from .importer import parseScene
 from .exporter.to_c import clearBootupScene, modifySceneTable, editSpecFile, deleteSceneFiles
